transações ok. Exceçoes bugado

+++ Pessoa.hpp
@@ -2,6 +2,7 @@
 #define PESSOA_HPP
 
 #include <string>
+#include <cstdint>
 
 
 class ContaBancaria;
@@ -10,26 +11,26 @@ class Pessoa{
     public:
         Pessoa();
 	    Pessoa(const std::string& nome, const unsigned short idade);
-        Pessoa(const std::string& nome, const unsigned short idade, const unsigned long cpf);
+        Pessoa(const std::string& nome, const unsigned short idade, const uint32_t cpf);
         virtual ~Pessoa();
 
 
         std::string getNome() const;
         unsigned short  getIdade() const;
-        unsigned long getCpf() const;
+        uint32_t getCpf() const;
 
         void setNome(const std::string& nome);
         void setIdade(const unsigned short idade);
-        void setCpf(unsigned long cpfTeste);
+        void setCpf(uint32_t cpfTeste);
 
-        bool validarCPF(unsigned long cpfTeste) const;
+        bool validarCPF(uint32_t cpfTeste) const;
 
 
 
     private:
         std::string nome;
         unsigned char idade;
-        unsigned long cpf;
+        uint32_t cpf;
 
 };
